[Servo] Move `enforcePositionLimits` and `enforceVelocityLimits` to utilities (#2180)
* Move limit enforcing functions to utilities

* Fix comments

* Make clock const

* Remove clock from enforcePositionLimit

* Remove clock usage from transformTwistToPlanningFrame and applyJointUpdates

* Remove clock from vvelocityScalingFactorForSingularity

* Fix tests

* Cleanups + clang-tidy

* Minor cleanups

* Log output formatting
