+++ moveit_ros/moveit_servo/include/moveit_servo/servo_calcs.h
@@ -169,11 +169,6 @@ protected:
   void suddenHalt(sensor_msgs::msg::JointState& joint_state,
                   const std::vector<const moveit::core::JointModel*>& joints_to_halt) const;
 
-  /** \brief Avoid overshooting joint limits
-      \return Vector of the joints that would move farther past position margin limits
-   */
-  std::vector<const moveit::core::JointModel*> enforcePositionLimits(sensor_msgs::msg::JointState& joint_state) const;
-
   /** \brief Compose the outgoing JointTrajectory message */
   void composeJointTrajMessage(const sensor_msgs::msg::JointState& joint_state,
                                trajectory_msgs::msg::JointTrajectory& joint_trajectory);
