+++ moveit_ros/moveit_servo/include/moveit_servo/utilities.h
@@ -61,7 +61,6 @@ geometry_msgs::msg::TransformStamped convertIsometryToTransform(const Eigen::Iso
  * @param[in] hard_stop_singularity_threshold  Halt if condition(Jacobian) > hard_stop_singularity_threshold
  * @param[in] lower_singularity_threshold      Decelerate if condition(Jacobian) > lower_singularity_threshold
  * @param[in] leaving_singularity_threshold_multiplier      Allow faster motion away from singularity
- * @param[in, out] clock          A ROS clock, for logging
  * @param[in, out] current_state  The state of the robot. Used in internal calculations.
  * @param[out] status             Singularity status
  */
@@ -71,12 +70,11 @@ double velocityScalingFactorForSingularity(const moveit::core::JointModelGroup* 
                                            const Eigen::MatrixXd& pseudo_inverse,
                                            const double hard_stop_singularity_threshold,
                                            const double lower_singularity_threshold,
-                                           const double leaving_singularity_threshold_multiplier, rclcpp::Clock& clock,
+                                           const double leaving_singularity_threshold_multiplier,
                                            const moveit::core::RobotStatePtr& current_state, StatusCode& status);
 
 /** \brief Joint-wise update of a sensor_msgs::msg::JointState with given delta's
  * Also filters and calculates the previous velocity
- * @param clock A ROS clock, for logging
  * @param publish_period The publishing rate for servo command
  * @param delta_theta Eigen vector of joint delta's
  * @param previous_joint_state The previous joint state
@@ -84,7 +82,7 @@ double velocityScalingFactorForSingularity(const moveit::core::JointModelGroup* 
  * @param smoother The trajectory smoother to be used.
  * @return Returns false if there is a problem, true otherwise
  */
-bool applyJointUpdate(rclcpp::Clock& clock, const double publish_period, const Eigen::ArrayXd& delta_theta,
+bool applyJointUpdate(const double publish_period, const Eigen::ArrayXd& delta_theta,
                       const sensor_msgs::msg::JointState& previous_joint_state,
                       sensor_msgs::msg::JointState& next_joint_state,
                       pluginlib::UniquePtr<online_signal_smoothing::SmoothingBaseClass>& smoother);
@@ -93,10 +91,9 @@ bool applyJointUpdate(rclcpp::Clock& clock, const double publish_period, const E
  * @param cmd The twist command received from the user
  * @param planning_frame Moveit planning frame of the robot
  * @param current_state The state of the robot
- * @param clock A ROS clock, for logging
  */
 void transformTwistToPlanningFrame(geometry_msgs::msg::TwistStamped& cmd, const std::string& planning_frame,
-                                   const moveit::core::RobotStatePtr& current_state, rclcpp::Clock& clock);
+                                   const moveit::core::RobotStatePtr& current_state);
 
 /** \brief Converts the delta_x (change in cartesian position) to a pose to be used with IK solver.
  * @param delta_x The change in cartesian position
@@ -105,4 +102,35 @@ void transformTwistToPlanningFrame(geometry_msgs::msg::TwistStamped& cmd, const 
  */
 geometry_msgs::msg::Pose poseFromCartesianDelta(const Eigen::VectorXd& delta_x,
                                                 const Eigen::Isometry3d& base_to_tip_frame_transform);
+
+/**
+ * @brief Computes the velocity scaling factor based on the joint velocity limits
+ * @param joint_model_group The MoveIt group
+ * @param velocity The vector containing the target velocities of the joints
+ * @return The velocity scaling factor
+ */
+double getVelocityScalingFactor(const moveit::core::JointModelGroup* joint_model_group, const Eigen::VectorXd& velocity);
+
+/**
+ * @brief Decrease robot position change and velocity, if needed, to satisfy joint velocity limits
+ * @param joint_model_group Active joint group. Used to retrieve limits
+ * @param publish_period Period of the servo loop
+ * @param joint_state The command that will go to the robot
+ * @param override_velocity_scaling_factor Option if the user wants a constant override of the velocity scaling.
+ *        A value greater than 0 will override the internal calculations done by getVelocityScalingFactor
+ */
+void enforceVelocityLimits(const moveit::core::JointModelGroup* joint_model_group, const double publish_period,
+                           sensor_msgs::msg::JointState& joint_state,
+                           const double override_velocity_scaling_factor = 0.0);
+
+/** \brief Avoid overshooting joint limits
+ * @param joint_state The joint state to be checked
+ * @param joint_limit_margin The allowed margin for joint limit. This is a buffer, prior to the actual limit.
+ * @param joint_model_group The MoveIt group
+ * @return Vector of the joints that would move farther past position margin limits
+ */
+std::vector<const moveit::core::JointModel*>
+enforcePositionLimits(sensor_msgs::msg::JointState& joint_state, const double joint_limit_margin,
+                      const moveit::core::JointModelGroup* joint_model_group);
+
 }  // namespace moveit_servo
